move: 파일 정리, style: top버튼 보이는 위치 수정
