+++ src/pages/home/index.tsx
@@ -2,7 +2,7 @@ import TopBtn from "../../components/common/TopBtn";
 import Landing from "./components/Landing";
 import OEIndex from "./components/OEIndex";
 import OeTip from "./components/OeTip";
-import Recipe from "./components/Recipe";
+import Recipe from "./components/recipe/Recipe";
 
 function Home() {
   return (
